Move EstonianLemmatizer object to __init__
