+++ spacy/lang/et/lemmatizer/lookup.py
@@ -7,8 +7,10 @@ from collections import defaultdict
 
 # https://raw.githubusercontent.com/michmech/lemmatization-lists/master/lemmatization-et.txt
 fnm = os.path.join(os.path.curdir(), 'dict.bz2')
-LOOKUP = defaultdict(list)
+lemmas = defaultdict(list)
 with bz2.BZ2File(fnm) as inf:
     for ln in inf:
         infl, lem = ln.decode('utf8').rstrip().split("\t", 1)
-        LOOKUP[infl].append(lem[0])
+        lemmas[infl].append(lem[0])
+
+LOOKUP = lemmas
